requirements.txt
# Core chess logic
python-chess>=1.999
numpy>=1.24

# CV / ML pipeline
opencv-contrib-python>=4.8
tensorflow>=2.13
scikit-learn>=1.3

# Optional but common for robotics serial interfaces (Hiwonder/LX-16A)
pyserial>=3.5
